Prototyping login for collection of question answer text and question answer images. Currently decided to not build out functionality for the automatic collection of question answers.
It takes noticeably longer to load the page when logged in.
There seems to be no efficient means for downloading a specific image directly to a folder.

Logging in works automatically. Can now log in and extract text as well, not blocked by perimeterX.

Header names for all elements are different when logged in. It is a different site, so to collect questions when logged in will require an extra method for the different element/class names.

Improved on the menu for the application. Added more features for adjusting the values of question scraping.

Added multiple settings files for umatrix.